add a pre-generated irods_wrap.c and precompiled irods.py
Running swig on this codebase is error-prone.
It seems you have to use the right swig version (2.0.10) and
then overwrite one function that swig generates wrongly for you.

I don't know where this functions comes from as the original
source also delivers a pre-generated irods_wrap.c

add a patch application in build.sh that replaces the function
with what it's suppposed to be. Now you should be able to run
./build.sh with a correct outcome at least if you have
swig 2.0.10.