copy the status of sendApiRequest into the conn status
General problem:
There is no way in python to find out if a connection is healthy
or not.

Observation:
The status variable of rcConn_t is never used in the whole
code base.
The status output of sendApiRequest concerns only connection-
related errors.

Fix:
copy the status of sendApiRequest into conn->status

This fix probably does not catch all the errors related to
irods connections, but should catch many, since the sendApiRequest
is used virtually everywhere.
For me  it fixed the problem that I was not able to detect a
connection error when the irodsAgent on the server failed.
My client is now able to detect these errors (as -5032) and
establish a new connection, which will then go to another
irodsAgent.